Add files via upload
A predictive model is developed from a dataset. Supervised machine learning techniques implemented in Python
implemented in Python language. For this purpose, a dataset of the New York Police Department for the years 2009 and 2010 is used. (2009_1perc.csv and 2010_1perc.csv) which will give us information about the police procedures performed.